Added type cast for clarity in some calculations

+++ ssn_CommandLineToArgvW.c
@@ -65,7 +65,7 @@ WCHAR** ssn_CommandLineToArgvW(const WCHAR* lpCmdLine, int* pNumArgs)
 			return 0;
 		}
 
-		size_t executable_name_length = (size_t)GetModuleFileNameW(0, (WCHAR*)((uintptr_t)executable_name_buffer + (2 * sizeof(WCHAR*))), (executable_name_buffer_size - (2 * sizeof(WCHAR*))) / sizeof(WCHAR));
+		size_t executable_name_length = (size_t)GetModuleFileNameW(0, (WCHAR*)((uintptr_t)executable_name_buffer + (2 * sizeof(WCHAR*))), (DWORD)((executable_name_buffer_size - (2 * sizeof(WCHAR*))) / sizeof(WCHAR)));
 		if (!executable_name_length || executable_name_length > (executable_name_buffer_size / sizeof(WCHAR)))
 		{
 			WCHAR** new_executable_name_buffer = (WCHAR**)LocalReAlloc(executable_name_buffer, (2 * sizeof(WCHAR*)) + ((max_file_name_length + 1) * sizeof(WCHAR)), 0);
@@ -86,7 +86,7 @@ WCHAR** ssn_CommandLineToArgvW(const WCHAR* lpCmdLine, int* pNumArgs)
 				return 0;
 			}
 
-			executable_name_length = (size_t)(size_t)GetModuleFileNameW(0, (WCHAR*)((uintptr_t)executable_name_buffer + (2 * sizeof(WCHAR*))), (executable_name_buffer_size - (2 * sizeof(WCHAR*))) / sizeof(WCHAR));
+			executable_name_length = (size_t)GetModuleFileNameW(0, (WCHAR*)((uintptr_t)executable_name_buffer + (2 * sizeof(WCHAR*))), (DWORD)((executable_name_buffer_size - (2 * sizeof(WCHAR*))) / sizeof(WCHAR)));
 			if (!executable_name_length || executable_name_length > (executable_name_buffer_size / sizeof(WCHAR)))
 			{
 				DWORD realloc_error = GetLastError();
